requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
supabase==2.10.0
sqlparse==0.5.1